% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxcdashboard-package.R
\docType{data}
\name{colunga}
\alias{colunga}
\title{Colunga}
\format{
\subsection{\code{colunga}}{

A data frame with 174 rows and 41 columns
}
}
\source{
U de los Andes?
}
\usage{
colunga
}
\description{
Explicar en que consiste esta tabla
}
\keyword{datasets}
