% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxcdashboard-package.R
\docType{data}
\name{datos1}
\alias{datos1}
\title{Datos 1}
\format{
\subsection{\code{datos1}}{

A vector with 20 elements
}
}
\source{
U de los Andes?
}
\usage{
datos1
}
\description{
Explicar en que consiste este vector
}
\keyword{datasets}
