% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxcdashboard-package.R
\docType{data}
\name{datos2}
\alias{datos2}
\title{Datos 2}
\format{
\subsection{\code{datos2}}{

A vector with 20 elements
}
}
\source{
U de los Andes?
}
\usage{
datos2
}
\description{
Explicar en que consiste este vector
}
\keyword{datasets}
