% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxcdashboard-package.R
\docType{data}
\name{datos3}
\alias{datos3}
\title{Datos 3}
\format{
\subsection{\code{datos3}}{

A vector with 20 elements
}
}
\source{
U de los Andes?
}
\usage{
datos3
}
\description{
Explicar en que consiste este vector
}
\keyword{datasets}
