% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxcdashboard-package.R
\docType{data}
\name{datos4}
\alias{datos4}
\title{Datos 4}
\format{
\subsection{\code{datos4}}{

A vector with 20 elements
}
}
\source{
U de los Andes?
}
\usage{
datos4
}
\description{
Explicar en que consiste este vector
}
\keyword{datasets}
