% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxcdashboard-package.R
\docType{data}
\name{datos5}
\alias{datos5}
\title{Datos 5}
\format{
\subsection{\code{datos5}}{

A vector with 20 elements
}
}
\source{
U de los Andes?
}
\usage{
datos5
}
\description{
Explicar en que consiste este vector
}
\keyword{datasets}
