% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxcdashboard-package.R
\docType{data}
\name{datos6}
\alias{datos6}
\title{Datos 6}
\format{
\subsection{\code{datos6}}{

A vector with 20 elements
}
}
\source{
U de los Andes?
}
\usage{
datos6
}
\description{
Explicar en que consiste este vector
}
\keyword{datasets}
