% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxcdashboard-package.R
\docType{data}
\name{donaciones_mallplaza}
\alias{donaciones_mallplaza}
\title{Donaciones Mallplaza}
\format{
\subsection{\code{donaciones_mallplaza}}{

A data frame with 13 rows and 11 columns
}
}
\source{
U de los Andes?
}
\usage{
donaciones_mallplaza
}
\description{
Explicar en que consiste esta tabla
}
\keyword{datasets}
