% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxcdashboard-package.R
\docType{data}
\name{gtrends}
\alias{gtrends}
\title{Gtrends}
\format{
\subsection{\code{gtrends}}{

A data frame with 28 rows and 6 columns
}
}
\source{
U de los Andes?
}
\usage{
gtrends
}
\description{
Explicar en que consiste esta tabla
}
\keyword{datasets}
