% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxcdashboard-package.R
\docType{data}
\name{metricas_comuna}
\alias{metricas_comuna}
\title{Metricas Comuna}
\format{
\subsection{\code{metricas_comuna}}{

A data frame with 11 rows and 3 columns
}
}
\source{
U de los Andes?
}
\usage{
metricas_comuna
}
\description{
Explicar en que consiste esta tabla
}
\keyword{datasets}
