% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxcdashboard-package.R
\docType{data}
\name{proyectos_2016}
\alias{proyectos_2016}
\title{Proyectos 2016}
\format{
\subsection{\code{proyectos_2016}}{

A data frame with 65 rows and 12 columns
}
}
\source{
U de los Andes?
}
\usage{
proyectos_2016
}
\description{
Explicar en que consiste esta tabla
}
\keyword{datasets}
