% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxcdashboard-package.R
\docType{data}
\name{tot_categorias}
\alias{tot_categorias}
\title{Tot Categorias}
\format{
\subsection{\code{tot_categorias}}{

A data frame with 10 rows and 2 columns
}
}
\source{
U de los Andes?
}
\usage{
tot_categorias
}
\description{
Explicar en que consiste esta tabla
}
\keyword{datasets}
