% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sxcdashboard-package.R
\docType{data}
\name{variables}
\alias{variables}
\title{Variables}
\format{
\subsection{\code{variables}}{

A vector with 21 elements
}
}
\source{
U de los Andes?
}
\usage{
variables
}
\description{
Explicar en que consiste esta tabla
}
\keyword{datasets}
